// src/app/core/models/user.ts
export interface User {
    uid: string;
    accountType: string;
    firstName: string;
    middleName: string;
    lastName: string;
    dob: string;
    email: string;
    address: string;
    mobilePhone: string;
    workPhone: string;
    rank: string;
    shiftPreference: string;
    dayOnePreference: string;
    dayTwoPreference: string;
    userVerified: boolean;
    photoURL: string;
 }